Enable LEA cipher support
LEA is a South Korean 128-bit block cipher included in the ISO/IEC
29192-2:2019 standard (Information security - Lightweight cryptography -
Part 2: Block ciphers). It performs faster than AES without the use of
dedicated crypto instructions. If SIMD instructions were supported, it
would show faster performance, but even then, it would not be a good
choice as it would not outperform the dedicated crypto instructions.
Therefore, only use LEA-256-XTS if there is a clear reason to do so.
Also, to enable LEA support, it needs to be enabled in the kernel crypto
API.

Add the following EncryptionOption:
LEA_256_XTS = 11
LEA_256_CTS = 12